Move all SyncTask to synctask package
