+++ sharding-scaling-core/src/main/java/info/avalon566/shardingscaling/core/synctask/SyncTask.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package info.avalon566.shardingscaling.core.job.sync;
+package info.avalon566.shardingscaling.core.synctask;
 
 import info.avalon566.shardingscaling.core.controller.SyncTaskProgress;
